feat(chat): merge chat screens
